requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.0
requests==2.31.0
nltk==3.8.1
transformers==4.30.0
beautifulsoup4==4.12.2
torch==2.1.1 -f https://download.pytorch.org/whl/cpu/torch_stable.html
sentencepiece==0.1.99
accelerate==0.23.0
datasets==2.14.5
seqeval==1.2.2